CORE-340: storage -> rspace
