+++ models/src/test/scala/coop/rchain/models/RhoTypesTest.scala
@@ -5,7 +5,7 @@ import org.scalatest.{FlatSpec, Matchers}
 import implicits._
 import testImplicits._
 import BitSetBytesMapper._
-import coop.rchain.storage.Serialize
+import coop.rchain.rspace.Serialize
 
 import scala.collection.immutable.BitSet
