--- storage/src/main/protobuf/StorageBlobs.proto
@@ -1,26 +1,0 @@
-syntax = "proto3";
-
-option java_package = "coop.rchain.storage.datamodels";
-
-message BytesList {
-    repeated bytes values = 1;
-}
-
-message PsKsBytes {
-    BytesList patterns = 1;
-    bytes kvalue = 2;
-    bool persist = 3;
-}
-
-message PsKsBytesList {
-    repeated PsKsBytes values = 1;
-}
-
-message AsBytes {
-    bytes avalue = 1;
-    bool persist = 2;
-}
-
-message AsBytesList {
-    repeated AsBytes values = 1;
-}
