--- storage/src/main/scala/coop/rchain/storage/Match.scala
@@ -1,12 +1,0 @@
-package coop.rchain.storage
-
-/**
-  * Type class for matching patterns with data.
-  *
-  * @tparam P A type representing patterns
-  * @tparam A A type representing data
-  */
-trait Match[P, A] {
-
-  def get(p: P, a: A): Option[A]
-}
